+++ warp/fem/geometry/grid_3d.py
@@ -328,7 +328,7 @@ class Grid3D(Geometry):
     def side_position(args: SideArg, s: Sample):
         side = Grid3D.get_side(args, s.element_index)
 
-        coord0 = wp.select(side.origin[0] == 0, s.element_coords[0], 1.0 - s.element_coords[0])
+        coord0 = wp.where(side.origin[0] == 0, 1.0 - s.element_coords[0], s.element_coords[0])
 
         local_pos = wp.vec3(
             float(side.origin[0]),
@@ -344,7 +344,7 @@ class Grid3D(Geometry):
     def side_deformation_gradient(args: SideArg, s: Sample):
         side = Grid3D.get_side(args, s.element_index)
 
-        sign = wp.select(side.origin[0] == 0, 1.0, -1.0)
+        sign = wp.where(side.origin[0] == 0, -1.0, 1.0)
 
         return wp.matrix_from_cols(
             wp.cw_mul(Grid3D._local_to_world(side.axis, wp.vec3(0.0, sign, 0.0)), args.cell_arg.cell_size),
@@ -376,7 +376,7 @@ class Grid3D(Geometry):
     def side_normal(args: SideArg, s: Sample):
         side = Grid3D.get_side(args, s.element_index)
 
-        sign = wp.select(side.origin[0] == 0, 1.0, -1.0)
+        sign = wp.where(side.origin[0] == 0, -1.0, 1.0)
 
         local_n = wp.vec3(sign, 0.0, 0.0)
         return Grid3D._local_to_world(side.axis, local_n)
@@ -385,7 +385,7 @@ class Grid3D(Geometry):
     def side_inner_cell_index(arg: SideArg, side_index: ElementIndex):
         side = Grid3D.get_side(arg, side_index)
 
-        inner_alt = wp.select(side.origin[0] == 0, side.origin[0] - 1, 0)
+        inner_alt = wp.where(side.origin[0] == 0, 0, side.origin[0] - 1)
 
         inner_origin = wp.vec3i(inner_alt, side.origin[1], side.origin[2])
 
@@ -398,8 +398,8 @@ class Grid3D(Geometry):
 
         alt_axis = Grid3D._local_to_world_axis(side.axis, 0)
 
-        outer_alt = wp.select(
-            side.origin[0] == arg.cell_arg.res[alt_axis], side.origin[0], arg.cell_arg.res[alt_axis] - 1
+        outer_alt = wp.where(
+            side.origin[0] == arg.cell_arg.res[alt_axis], arg.cell_arg.res[alt_axis] - 1, side.origin[0]
         )
 
         outer_origin = wp.vec3i(outer_alt, side.origin[1], side.origin[2])
@@ -411,9 +411,9 @@ class Grid3D(Geometry):
     def side_inner_cell_coords(args: SideArg, side_index: ElementIndex, side_coords: Coords):
         side = Grid3D.get_side(args, side_index)
 
-        inner_alt = wp.select(side.origin[0] == 0, 1.0, 0.0)
+        inner_alt = wp.where(side.origin[0] == 0, 0.0, 1.0)
 
-        side_coord0 = wp.select(side.origin[0] == 0, side_coords[0], 1.0 - side_coords[0])
+        side_coord0 = wp.where(side.origin[0] == 0, 1.0 - side_coords[0], side_coords[0])
 
         return Grid3D._local_to_world(side.axis, wp.vec3(inner_alt, side_coord0, side_coords[1]))
 
@@ -422,9 +422,9 @@ class Grid3D(Geometry):
         side = Grid3D.get_side(args, side_index)
 
         alt_axis = Grid3D._local_to_world_axis(side.axis, 0)
-        outer_alt = wp.select(side.origin[0] == args.cell_arg.res[alt_axis], 0.0, 1.0)
+        outer_alt = wp.where(side.origin[0] == args.cell_arg.res[alt_axis], 1.0, 0.0)
 
-        side_coord0 = wp.select(side.origin[0] == 0, side_coords[0], 1.0 - side_coords[0])
+        side_coord0 = wp.where(side.origin[0] == 0, 1.0 - side_coords[0], side_coords[0])
 
         return Grid3D._local_to_world(side.axis, wp.vec3(outer_alt, side_coord0, side_coords[1]))
 
@@ -442,7 +442,7 @@ class Grid3D(Geometry):
             long_axis = Grid3D._local_to_world_axis(side.axis, 1)
             lat_axis = Grid3D._local_to_world_axis(side.axis, 2)
             long_coord = element_coords[long_axis]
-            long_coord = wp.select(side.origin[0] == 0, long_coord, 1.0 - long_coord)
+            long_coord = wp.where(side.origin[0] == 0, 1.0 - long_coord, long_coord)
             return Coords(long_coord, element_coords[lat_axis], 0.0)
 
         return Coords(OUTSIDE)